Added nj_eitc_income_eligible variable
